@@ -3,11 +3,12 @@
   "version": "1.0.0",
   "description": "curriculum-vitae",
   "main": "entry.js",
+  "type": "module",
   "scripts": {
     "start": "yarn prod && docker build -f Dockerfile -t cv-pwahost . && docker run --name cv-pwahost --rm -d -p 8080:80/tcp cv-pwahost:latest",
-    "dev": "yarn build && webpack serve --mode development --config ./webpack.dev.js",
-    "prod": "yarn clean && set NODE_ENV=production && webpack --mode production --config ./webpack.prod.js",
-    "build": "yarn clean && set NODE_ENV=development && webpack --mode development --config ./webpack.dev.js",
+    "dev": "yarn build && webpack serve --mode development --config ./webpack.dev-config.js",
+    "prod": "yarn clean && set NODE_ENV=production && webpack --mode production --config ./webpack.prod-config.js",
+    "build": "yarn clean && set NODE_ENV=development && webpack --mode development --config ./webpack.dev-config.js",
     "clean": "rm -rf dist",
     "deploy": "yarn prod && firebase deploy",
     "deploy-cloud": "yarn prod && firebase deploy --debug --token \"$FIREBASE_TOKEN\"",
@@ -17,40 +18,37 @@
   "author": "Y.Goalen",
   "license": "MIT",
   "devDependencies": {
-    "base64-inline-loader": "^1.1.1",
-    "browser-sync": "^2.26.12",
-    "browser-sync-webpack-plugin": "^2.2.2",
-    "copy-webpack-plugin": "^6.0.3",
-    "css-loader": "^2.1.0",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "file-loader": "^3.0.1",
-    "html-loader": "^1.3.0",
+    "base64-inline-loader": "^2.0.1",
+    "browser-sync": "^2.27.10",
+    "browser-sync-webpack-plugin": "^2.3.0",
+    "copy-webpack-plugin": "^11.0.0",
+    "css-loader": "^6.7.1",
+    "css-minimizer-webpack-plugin": "^4.0.0",
+    "file-loader": "^6.2.0",
+    "html-loader": "^3.1.0",
     "html-loader-srcset": "^0.5.6",
     "html-srcsets-loader": "^0.5.1",
-    "html-webpack-plugin": "^4.4.1",
-    "image-webpack-loader": "^7.0.1",
-    "imagemin": "^7.0.1",
-    "imagemin-webp": "^6.0.0",
-    "mini-css-extract-plugin": "^0.5.0",
-    "node-sass": "^4.11.0",
+    "html-webpack-plugin": "^5.5.0",
+    "image-webpack-loader": "^8.1.0",
+    "imagemin": "^8.0.1",
+    "imagemin-webp": "^7.0.0",
+    "mini-css-extract-plugin": "^2.6.0",
+    "node-sass": "^7.0.1",
     "offline-plugin": "^5.0.7",
     "popper.js": "^1.14.7",
-    "resolve-url-loader": "^3.1.1",
-    "sass-loader": "^7.1.0",
-    "style-loader": "^0.23.1",
-    "svg-sprite-loader": "^5.2.1",
-    "svg-spritemap-webpack-plugin": "^3.8.2",
-    "svg-url-loader": "^6.0.0",
-    "url-loader": "^4.1.0",
+    "resolve-url-loader": "^5.0.0",
+    "sass-loader": "^13.0.0",
+    "style-loader": "^3.3.1",
+    "svg-sprite-loader": "^6.0.11",
+    "svg-spritemap-webpack-plugin": "^4.4.0",
+    "svg-url-loader": "^7.1.1",
+    "url-loader": "^4.1.1",
     "webp-loader": "^0.6.0",
-    "webpack": "^4.29.6",
-    "webpack-bundle-analyzer": "^3.8.0",
-    "webpack-cli": "^4.6.0",
+    "webpack": "^5.73.0",
+    "webpack-bundle-analyzer": "^4.5.0",
+    "webpack-cli": "^4.9.2",
     "webpack-css-replace-images-to-webp": "^1.0.1",
-    "webpack-dev-server": "^3.11.2",
-    "webpack-merge": "^5.0.9"
-  },
-  "dependencies": {
-    "optimize-css-assets-webpack-plugin": "^5.0.4"
+    "webpack-dev-server": "^4.9.1",
+    "webpack-merge": "^5.8.0"
   }
 }
